Description
Objective of Driver Behavior Monitoring System is to detect the combined state of driver drowsiness, fatigue and inattention while driving and alert the driver before they move to state where they won’t be able to react. Human life is greatly affected by roadside accidents as people lose their control on certain situations while driving. This can be prevented by continuously monitoring the driver .behavior while driving and alert them on right time. This study focuses on detecting the driver behavior, so that the driver can be alerted on time. Previous researchers have worked on various metrics to monitor the driver behavior related to fatigue while driving. This study uses camera-based monitoring of driver behavior on detecting the face, eyes and mouth of the driver. Viola Jones, skin segmentation and SVM classifier algorithm are used in face, eyes and mouth detection from live video. Based on the detection of eyes and mouth, the SVM classifier is trained and tested to provide correct detection. This system is a successful non-intrusive way of early detection on fatigue/drowsiness state while driving. In future, this GUI system can be connected to android application on phone to alert the driver before they move on to the state where they are unable to react.